+++ src/main/java/com/loxil/helpstarter/MinecraftUtils.java
@@ -1,7 +1,12 @@
 package com.loxil.helpstarter;
 
 import net.minecraft.client.Minecraft;
+import net.minecraft.client.network.NetworkPlayerInfo;
 import net.minecraft.util.ChatComponentText;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
 
 public final class MinecraftUtils {
 
@@ -10,6 +15,17 @@ public final class MinecraftUtils {
 
     public static String getUsername() {
         return Minecraft.getMinecraft().getSession().getUsername();
+    }
+
+    public static List<String> getOnlinePlayerNames() {
+        List<String> playerNames = new ArrayList<String>();
+        if (Minecraft.getMinecraft().thePlayer != null && Minecraft.getMinecraft().thePlayer.sendQueue != null) {
+            Collection<NetworkPlayerInfo> playerInfoList = Minecraft.getMinecraft().thePlayer.sendQueue.getPlayerInfoMap();
+            for (NetworkPlayerInfo playerInfo : playerInfoList) {
+                playerNames.add(playerInfo.getGameProfile().getName());
+            }
+        }
+        return playerNames;
     }
 
     public static void addChatMessageOnMainThread(final String message) {